Use setters instead of requirng every field
