+++ cmd/kubeapps-apis/plugin_test/mock_plugin.go
@@ -22,39 +22,9 @@ import (
 	"google.golang.org/grpc/status"
 )
 
-const (
-	GlobalPackagingCluster  = "default"
-	DefaultAppVersion       = "1.2.6"
-	DefaultCategory         = "cat-1"
-	DefaultIconURL          = "https://example.com/package.svg"
-	DefaultHomeURL          = "https://example.com/home"
-	DefaultRepoURL          = "https://example.com/repo"
-	DefaultDescription      = "description"
-	DefaultId               = "repo-1/package-id"
-	DefaultNamespace        = "my-namespace-1"
-	DefaultPkgVersion       = "1.0.0"
-	DefaultPkgUpdateVersion = "2.0.0"
-	DefaultReleaseName      = "my-release-1"
-	DefaultReleaseNamespace = "my-release-namespace-1"
-	DefaultReleaseNotes     = "some notes"
-	DefaultReleaseValues    = "{\"value\":\"new\"}"
-	DefaultReleaseVersion   = "1.2.3"
-	DefaultValuesSchema     = "\"$schema\": \"http://json-schema.org/schema#\""
-	DefaultReadme           = "#readme"
-	DefaultValues           = "key: value"
-	DefaultMaintainerName   = "me"
-	DefaultMaintainerEmail  = "me@example.com"
-)
-
-var defaultInstalledPackageStatus = &corev1.InstalledPackageStatus{
-	Ready:      true,
-	Reason:     corev1.InstalledPackageStatus_STATUS_REASON_INSTALLED,
-	UserReason: "ReconciliationSucceeded",
-}
-
-type TestPackagingPlugin struct {
+type TestPackagingPluginServer struct {
 	packages.UnimplementedPackagesServiceServer
-	plugin                    *plugins.Plugin
+	Plugin                    *plugins.Plugin
 	availablePackageSummaries []*corev1.AvailablePackageSummary
 	availablePackageDetail    *corev1.AvailablePackageDetail
 	installedPackageSummaries []*corev1.InstalledPackageSummary
@@ -65,162 +35,39 @@ type TestPackagingPlugin struct {
 	status                    codes.Code
 }
 
-type MockPkgsPluginWithServer struct {
-	Plugin *plugins.Plugin
-	Server packages.PackagesServiceServer
+func (s *TestPackagingPluginServer) SetAvailablePackageSummaries(availablePackageSummaries []*corev1.AvailablePackageSummary) {
+	s.availablePackageSummaries = availablePackageSummaries
+}
+func (s *TestPackagingPluginServer) SetAvailablePackageDetail(availablePackageDetail *corev1.AvailablePackageDetail) {
+	s.availablePackageDetail = availablePackageDetail
+}
+func (s *TestPackagingPluginServer) SetInstalledPackageSummary(installedPackageSummaries []*corev1.InstalledPackageSummary) {
+	s.installedPackageSummaries = installedPackageSummaries
+}
+func (s *TestPackagingPluginServer) SetInstalledPackageDetail(installedPackageDetail *corev1.InstalledPackageDetail) {
+	s.installedPackageDetail = installedPackageDetail
+}
+func (s *TestPackagingPluginServer) SetPackageAppVersion(packageAppVersions []*corev1.PackageAppVersion) {
+	s.packageAppVersions = packageAppVersions
+}
+func (s *TestPackagingPluginServer) SetCategories(categories []string) {
+	s.categories = categories
+}
+func (s *TestPackagingPluginServer) SetNextPageToken(nextPageToken string) {
+	s.nextPageToken = nextPageToken
+}
+func (s *TestPackagingPluginServer) SetStatus(status codes.Code) {
+	s.status = status
 }
 
-func NewTestPackagingPlugin(plugin *plugins.Plugin) *TestPackagingPlugin {
-	return &TestPackagingPlugin{
-		plugin: plugin,
-	}
-}
-
-func MakeTestPackagingPlugin(
-	plugin *plugins.Plugin,
-	availablePackageSummaries []*corev1.AvailablePackageSummary,
-	availablePackageDetail *corev1.AvailablePackageDetail,
-	installedPackageSummaries []*corev1.InstalledPackageSummary,
-	installedPackageDetail *corev1.InstalledPackageDetail,
-	packageAppVersions []*corev1.PackageAppVersion,
-	nextPageToken string,
-	categories []string,
-	status codes.Code,
-) *MockPkgsPluginWithServer {
-	return &MockPkgsPluginWithServer{
+func NewTestPackagingPlugin(plugin *plugins.Plugin) *TestPackagingPluginServer {
+	return &TestPackagingPluginServer{
 		Plugin: plugin,
-		Server: &TestPackagingPlugin{
-			plugin:                    plugin,
-			availablePackageSummaries: availablePackageSummaries,
-			availablePackageDetail:    availablePackageDetail,
-			installedPackageSummaries: installedPackageSummaries,
-			installedPackageDetail:    installedPackageDetail,
-			packageAppVersions:        packageAppVersions,
-			categories:                categories,
-			nextPageToken:             nextPageToken,
-			status:                    status,
-		},
-	}
-}
-
-func MakeAvailablePackageSummary(name string, plugin *plugins.Plugin) *corev1.AvailablePackageSummary {
-	return &corev1.AvailablePackageSummary{
-		Name:        name,
-		DisplayName: name,
-		LatestVersion: &corev1.PackageAppVersion{
-			PkgVersion: DefaultPkgVersion,
-			AppVersion: DefaultAppVersion,
-		},
-		IconUrl:          DefaultIconURL,
-		Categories:       []string{DefaultCategory},
-		ShortDescription: DefaultDescription,
-		AvailablePackageRef: &corev1.AvailablePackageReference{
-			Context:    &corev1.Context{Cluster: GlobalPackagingCluster, Namespace: DefaultNamespace},
-			Identifier: DefaultId,
-			Plugin:     plugin,
-		},
-	}
-}
-
-func MakeInstalledPackageSummary(name string, plugin *plugins.Plugin) *corev1.InstalledPackageSummary {
-	return &corev1.InstalledPackageSummary{
-		InstalledPackageRef: &corev1.InstalledPackageReference{
-			Context: &corev1.Context{
-				Namespace: DefaultNamespace,
-			},
-			Identifier: name,
-			Plugin:     plugin,
-		},
-		Name:    name,
-		IconUrl: DefaultIconURL,
-		PkgVersionReference: &corev1.VersionReference{
-			Version: DefaultPkgVersion,
-		},
-		CurrentVersion: &corev1.PackageAppVersion{
-			PkgVersion: DefaultPkgVersion,
-			AppVersion: DefaultAppVersion,
-		},
-		PkgDisplayName:   name,
-		ShortDescription: DefaultDescription,
-		Status:           defaultInstalledPackageStatus,
-		LatestVersion: &corev1.PackageAppVersion{
-			PkgVersion: DefaultPkgVersion,
-			AppVersion: DefaultAppVersion,
-		},
-	}
-}
-
-func MakeAvailablePackageDetail(name string, plugin *plugins.Plugin) *corev1.AvailablePackageDetail {
-	return &corev1.AvailablePackageDetail{
-		AvailablePackageRef: &corev1.AvailablePackageReference{
-			Context:    &corev1.Context{Cluster: GlobalPackagingCluster, Namespace: DefaultNamespace},
-			Identifier: DefaultId,
-			Plugin:     plugin,
-		},
-		Name: name,
-		Version: &corev1.PackageAppVersion{
-			PkgVersion: DefaultPkgVersion,
-			AppVersion: DefaultAppVersion,
-		},
-		RepoUrl:          DefaultRepoURL,
-		IconUrl:          DefaultIconURL,
-		HomeUrl:          DefaultHomeURL,
-		DisplayName:      name,
-		Categories:       []string{DefaultCategory},
-		ShortDescription: DefaultDescription,
-		Readme:           DefaultReadme,
-		DefaultValues:    DefaultValues,
-		ValuesSchema:     DefaultValuesSchema,
-		SourceUrls:       []string{DefaultHomeURL},
-		Maintainers:      []*corev1.Maintainer{{Name: DefaultMaintainerName, Email: DefaultMaintainerEmail}},
-	}
-}
-
-func MakeInstalledPackageDetail(name string, plugin *plugins.Plugin) *corev1.InstalledPackageDetail {
-	return &corev1.InstalledPackageDetail{
-		InstalledPackageRef: &corev1.InstalledPackageReference{
-			Context: &corev1.Context{
-				Namespace: DefaultReleaseNamespace,
-				Cluster:   GlobalPackagingCluster,
-			},
-			Identifier: DefaultReleaseName,
-		},
-		PkgVersionReference: &corev1.VersionReference{
-			Version: DefaultReleaseVersion,
-		},
-		Name: DefaultReleaseName,
-		CurrentVersion: &corev1.PackageAppVersion{
-			PkgVersion: DefaultReleaseVersion,
-			AppVersion: DefaultAppVersion,
-		},
-		LatestVersion: &corev1.PackageAppVersion{
-			PkgVersion: DefaultReleaseVersion,
-			AppVersion: DefaultAppVersion,
-		},
-		ValuesApplied:         DefaultReleaseValues,
-		PostInstallationNotes: DefaultReleaseNotes,
-		Status:                defaultInstalledPackageStatus,
-		AvailablePackageRef: &corev1.AvailablePackageReference{
-			Context: &corev1.Context{
-				Namespace: DefaultReleaseNamespace,
-				Cluster:   GlobalPackagingCluster,
-			},
-			Identifier: DefaultId,
-			Plugin:     plugin,
-		},
-		CustomDetail: nil,
-	}
-}
-
-func MakePackageAppVersion(appVersion, pkgVersion string) *corev1.PackageAppVersion {
-	return &corev1.PackageAppVersion{
-		AppVersion: appVersion,
-		PkgVersion: pkgVersion,
 	}
 }
 
 // GetAvailablePackages returns the packages based on the request.
-func (s TestPackagingPlugin) GetAvailablePackageSummaries(ctx context.Context, request *packages.GetAvailablePackageSummariesRequest) (*packages.GetAvailablePackageSummariesResponse, error) {
+func (s TestPackagingPluginServer) GetAvailablePackageSummaries(ctx context.Context, request *packages.GetAvailablePackageSummariesRequest) (*packages.GetAvailablePackageSummariesResponse, error) {
 	if s.status != codes.OK {
 		return nil, status.Errorf(s.status, "Non-OK response")
 	}
@@ -232,7 +79,7 @@ func (s TestPackagingPlugin) GetAvailablePackageSummaries(ctx context.Context, r
 }
 
 // GetAvailablePackageDetail returns the package details based on the request.
-func (s TestPackagingPlugin) GetAvailablePackageDetail(ctx context.Context, request *packages.GetAvailablePackageDetailRequest) (*packages.GetAvailablePackageDetailResponse, error) {
+func (s TestPackagingPluginServer) GetAvailablePackageDetail(ctx context.Context, request *packages.GetAvailablePackageDetailRequest) (*packages.GetAvailablePackageDetailResponse, error) {
 	if s.status != codes.OK {
 		return nil, status.Errorf(s.status, "Non-OK response")
 	}
@@ -242,7 +89,7 @@ func (s TestPackagingPlugin) GetAvailablePackageDetail(ctx context.Context, requ
 }
 
 // GetInstalledPackageSummaries returns the installed package summaries based on the request.
-func (s TestPackagingPlugin) GetInstalledPackageSummaries(ctx context.Context, request *packages.GetInstalledPackageSummariesRequest) (*packages.GetInstalledPackageSummariesResponse, error) {
+func (s TestPackagingPluginServer) GetInstalledPackageSummaries(ctx context.Context, request *packages.GetInstalledPackageSummariesRequest) (*packages.GetInstalledPackageSummariesResponse, error) {
 	if s.status != codes.OK {
 		return nil, status.Errorf(s.status, "Non-OK response")
 	}
@@ -253,7 +100,7 @@ func (s TestPackagingPlugin) GetInstalledPackageSummaries(ctx context.Context, r
 }
 
 // GetInstalledPackageDetail returns the package versions based on the request.
-func (s TestPackagingPlugin) GetInstalledPackageDetail(ctx context.Context, request *packages.GetInstalledPackageDetailRequest) (*packages.GetInstalledPackageDetailResponse, error) {
+func (s TestPackagingPluginServer) GetInstalledPackageDetail(ctx context.Context, request *packages.GetInstalledPackageDetailRequest) (*packages.GetInstalledPackageDetailResponse, error) {
 	if s.status != codes.OK {
 		return nil, status.Errorf(s.status, "Non-OK response")
 	}
@@ -263,7 +110,7 @@ func (s TestPackagingPlugin) GetInstalledPackageDetail(ctx context.Context, requ
 }
 
 // GetAvailablePackageVersions returns the package versions based on the request.
-func (s TestPackagingPlugin) GetAvailablePackageVersions(ctx context.Context, request *packages.GetAvailablePackageVersionsRequest) (*packages.GetAvailablePackageVersionsResponse, error) {
+func (s TestPackagingPluginServer) GetAvailablePackageVersions(ctx context.Context, request *packages.GetAvailablePackageVersionsRequest) (*packages.GetAvailablePackageVersionsResponse, error) {
 	if s.status != codes.OK {
 		return nil, status.Errorf(s.status, "Non-OK response")
 	}
